Fix solving member selection type.
When the selected member's type is visible to the scope that defines the member,
but not to the scope of the member completion expression, the type of the
member cannot be solved, causing no completion. This is due to the type is not
solved using the scope that defines the member.