Use canonical name for solving imported classes.
JLS 8 section 7.5 specifies that import declarations uses canonical name.
Canonical name is defined in section 6.7.

If fully qualified name is used, solving super class and interfaces of a class
that is in part of one of the import declarations can cause infinite loop.
See TestClass.java in this commit for more details.