Distinguish between instatnce and non-instance members.
Entity instances how have a new method isStatic() based on whether
the "static" modifier is applied to an entity. As a special case,
top level classes are always considered is static. Entities that are
not classes, methods, or variables are always static. AstScanner
populates isStatic to class, method, and variable entities.

A new method isInstanceContext() is added to EntityWithContext so that
ExpressionSolver and TypeSolver knows whether or not instance members
should be included. Only methods and variables that are not static
are instance members.

IndexStore has been updated to serialize and deserialize isStatic
propoerty for classes, methods, and variables.

The following changes are not included in this commit:
 - Update JDK type index to include isStatic field.
 - Update completor to stop listing instance members in non-instance
   context.