Do not replace type variables with solved types in type index.
The purpose of solving type references is to get its full name so that
it can be solved with import statements removed from type indeces.

Type variables are not affected by the removal of type indeces. Solving
them prematurelly in type indeces only binds them to their upper bounds,
which is incorrect in most of the cases.