Unify the logic of shadowing entities with the same name.
Created a new class EntityShadowingListBuilder, which implements the logic
of shadowing new entities with existing ones. New entities are shadowed if there
is an existing entity that neither the new one nor the existing one is a method.

Methods that overrides each other should also be shadowed. This is not
implemented yet.