Find definition of constructors on new class statement.
Expression solver now returns the constructors of new class nodes with the
constructors of the class, prioritized using the argument matching result,
the same as method prioritization. If the class doesn't have any constructor,
the definition is the class itself.

Definition solver now recongizes new class expression and uses it to find
definitions.