Fix a parsing .class files with long or double constants.
Per JVMS 4.4.5: "All 8-byte constants take up two entries in the constant_pool
table of the class file. If a CONSTANT_Long_info or CONSTANT_Double_info
structure is the item in the constant_pool table at index n, then the next
usable item in the pool is located at index n+2. The constant_pool index n+1
must be valid but is considered unusable."

For long or double constants, we add them to the constant pool twice.