Improve fixing contents.
Handle two consecutive dots. The scanner returns error token instead of two
dots. Insert a dumb identifier between the two dots if the user is likely to
complete memeber selection between them.

In addition to that, do not insert semicolon after a dumb identifer for memeber
selection unless it's necessary.